Refactor proof handling and remove unused components
- Removed `SerializableProof` and `SerializableDecommitment` structs from `proof.rs` to streamline the codebase.
- Eliminated the `to_serializable_proof` function, simplifying proof serialization.
- Updated `lib.rs` to reflect the removal of unused exports and improve organization.
- Refactored `MaskedWitnessVarRevised` to `MaskedWitnessVar` and introduced a new `ValidationVar` for better clarity in validation processes.
- Consolidated test cases and removed outdated references to enhance maintainability.
