+++ schmivitz-snark/src/gadgets/masked_witness.rs
@@ -3,9 +3,9 @@ use ark_r1cs_std::boolean::Boolean;
 use ark_relations::r1cs::SynthesisError;
 use schmivitz::parameters::REPETITION_PARAM;
 
-pub struct MaskedWitnessVarRevised;
+pub struct MaskedWitnessVar;
 
-impl MaskedWitnessVarRevised {
+impl MaskedWitnessVar {
     /// Step 1: Compute d_delta values based on witness commitment and verifier key using boolean arrays
     #[tracing::instrument(
         target = "r1cs",
@@ -169,207 +169,11 @@ impl MaskedWitnessVarRevised {
 
         Ok(masked_witnesses)
     }
-
-    /// Step 3: Compute validation mask from mask voles using boolean arrays
-    ///
-    /// This corresponds to the calculation in proof.rs line 273:
-    #[tracing::instrument(target = "r1cs", skip(mask_voles_booleans))]
-    pub fn combine(
-        mask_voles_booleans: &Vec<Vec<Boolean<Bn254Fr>>>,
-    ) -> Result<Vec<Boolean<Bn254Fr>>, SynthesisError> {
-        // This function implements the combine function from proof.rs:
-
-        // Initialize accumulator with zeros (F128b::ZERO)
-        let mut acc_bits = Vec::with_capacity(128);
-        for _ in 0..128 {
-            acc_bits.push(Boolean::constant(false));
-        }
-
-        // Initialize power with ONE (F128b::ONE)
-        let mut power_bits = Vec::with_capacity(128);
-        power_bits.push(Boolean::constant(true)); // First bit is 1
-        for _ in 1..128 {
-            power_bits.push(Boolean::constant(false)); // Rest are 0
-        }
-
-        // Process each mask_vole value
-        for mask_vole_bits in mask_voles_booleans.iter() {
-            // Compute vi * power
-            // In boolean logic, multiplication is a bit-by-bit AND followed by appropriate shifts and XORs
-            // This is a simplified implementation that works for our specific case
-            let mut product_bits = Vec::with_capacity(128);
-            for _ in 0..128 {
-                product_bits.push(Boolean::constant(false));
-            }
-
-            // For each bit in mask_vole_bits
-            for (i, mask_bit) in mask_vole_bits.iter().enumerate() {
-                if i >= 128 {
-                    break;
-                }
-
-                // For each bit in power_bits
-                for (j, power_bit) in power_bits.iter().enumerate() {
-                    if j >= 128 || i + j >= 128 {
-                        continue;
-                    }
-
-                    // Compute mask_bit AND power_bit
-                    let and_result = Boolean::and(mask_bit, power_bit)?;
-
-                    // XOR with the corresponding bit in product_bits (at position i+j)
-                    product_bits[i + j] = Boolean::xor(&product_bits[i + j], &and_result)?;
-                }
-            }
-
-            // Add product to accumulator (XOR for boolean addition in F2)
-            for i in 0..128 {
-                acc_bits[i] = Boolean::xor(&acc_bits[i], &product_bits[i])?;
-            }
-        }
-        println!("mask_voles_booleans");
-
-        Ok(acc_bits)
-    }
-
-    /// Combined function to compute masked witnesses from witness commitment, verifier key, and witness voles
-    /// using boolean arrays
-    ///
-    /// This function combines the first two steps:
-    /// 1. Compute d_delta from witness commitment and verifier key
-    /// 2. Compute masked witnesses from witness voles and d_delta
-    #[tracing::instrument(
-        target = "r1cs",
-        skip(
-            witness_commitment_booleans,
-            verifier_key_booleans,
-            witness_voles_booleans
-        )
-    )]
-    pub fn compute(
-        witness_commitment_booleans: &Vec<Vec<Boolean<Bn254Fr>>>,
-        verifier_key_booleans: &Vec<Boolean<Bn254Fr>>,
-        witness_voles_booleans: &Vec<Vec<Vec<Boolean<Bn254Fr>>>>,
-    ) -> Result<Vec<Vec<Boolean<Bn254Fr>>>, SynthesisError> {
-        // Step 1: Compute d_delta
-        let d_delta_booleans =
-            Self::compute_d_delta(witness_commitment_booleans, verifier_key_booleans)?;
-
-        // Step 2: Compute masked witnesses
-        let masked_witnesses_booleans =
-            Self::compute_masked_witness(witness_voles_booleans, &d_delta_booleans)?;
-
-        Ok(masked_witnesses_booleans)
-    }
-    /// Compute validation aggregate directly using Boolean arrays
-    ///
-    /// This function replaces the need to convert Boolean arrays to FpVars for validation aggregate calculation
-    /// It implements the same logic as CircuitTraverser::compute_validation_aggregate but using Boolean operations
-    ///
-    /// # Arguments
-    ///
-    /// * `witness_challenges` - Array of witness challenges as Boolean arrays
-    /// * `masked_witnesses_var` - Array of masked witnesses as Boolean arrays
-    ///
-    /// # Returns
-    ///
-    /// * Result containing the validation aggregate as a Boolean array or a synthesis error
-    #[tracing::instrument(target = "r1cs", skip(witness_challenges, masked_witnesses_var))]
-    pub fn compute_validation_aggregate_revise(
-        witness_challenges: &[Vec<Boolean<Bn254Fr>>],
-        masked_witnesses_var: &[Vec<Boolean<Bn254Fr>>],
-    ) -> Result<Vec<Boolean<Bn254Fr>>, SynthesisError> {
-        if witness_challenges.len() > masked_witnesses_var.len() {
-            return Err(SynthesisError::Unsatisfiable);
-        }
-
-        // Initialize the validation aggregate with zeros (representing 0 in the field)
-        let mut validation_aggregate = Vec::with_capacity(128);
-        for _ in 0..128 {
-            validation_aggregate.push(Boolean::constant(false));
-        }
-
-        // Compute the validation aggregate as the sum of (challenge * masked_witness)
-        // In Boolean logic, multiplication is bit-by-bit AND followed by appropriate shifts and XORs
-        // Addition is XOR for the binary field
-        for (challenge, masked_witness) in
-            witness_challenges.iter().zip(masked_witnesses_var.iter())
-        {
-            // Compute challenge * masked_witness
-            let mut product = Vec::with_capacity(128);
-            for _ in 0..128 {
-                product.push(Boolean::constant(false));
-            }
-
-            // For each bit in challenge
-            for (i, challenge_bit) in challenge.iter().enumerate() {
-                if i >= 128 {
-                    break;
-                }
-
-                // For each bit in masked_witness
-                for (j, masked_bit) in masked_witness.iter().enumerate() {
-                    if j >= 128 || i + j >= 128 {
-                        continue;
-                    }
-
-                    // Compute challenge_bit AND masked_bit
-                    let and_result = Boolean::and(challenge_bit, masked_bit)?;
-
-                    // XOR with the corresponding bit in product (at position i+j)
-                    product[i + j] = Boolean::xor(&product[i + j], &and_result)?;
-                }
-            }
-
-            // Add product to validation_aggregate (XOR for boolean addition in F2)
-            for i in 0..128 {
-                validation_aggregate[i] = Boolean::xor(&validation_aggregate[i], &product[i])?;
-            }
-        }
-
-        Ok(validation_aggregate)
-    }
 }
 
 #[cfg(test)]
-mod revised_tests {
+mod tests {
     use super::*;
-    use ark_r1cs_std::R1CSVar;
-
-    #[test]
-    fn test_compute_validation_aggregate_revise() {
-        // Create test witness challenges (Boolean arrays representing field elements)
-        let witness_challenges = vec![
-            vec![Boolean::constant(true), Boolean::constant(false)], // Represents 1
-            vec![Boolean::constant(false), Boolean::constant(true)], // Represents 2
-        ];
-
-        // Create test masked witnesses (Boolean arrays representing field elements)
-        let masked_witnesses_var = vec![
-            vec![Boolean::constant(true), Boolean::constant(true)], // Represents 3
-            vec![Boolean::constant(false), Boolean::constant(false)], // Represents 0
-        ];
-
-        // Compute validation aggregate
-        let validation_aggregate = MaskedWitnessVarRevised::compute_validation_aggregate_revise(
-            &witness_challenges,
-            &masked_witnesses_var,
-        )
-        .unwrap();
-
-        // Verify the result has the expected length
-        assert_eq!(validation_aggregate.len(), 128);
-
-        // In this simple test case with 2-bit values:
-        // 1 * 3 + 2 * 0 = 3
-        // In binary: 11
-        // So the first two bits should be true, and the rest should be false
-        assert_eq!(validation_aggregate[0].value().unwrap(), true);
-        assert_eq!(validation_aggregate[1].value().unwrap(), true);
-        for i in 2..128 {
-            assert_eq!(validation_aggregate[i].value().unwrap(), false);
-        }
-    }
 
     #[test]
     fn test_compute_d_delta() {
@@ -383,11 +187,9 @@ mod revised_tests {
         let verifier_key_booleans = vec![Boolean::constant(true), Boolean::constant(true)];
 
         // Compute d_delta
-        let d_delta = MaskedWitnessVarRevised::compute_d_delta(
-            &witness_commitment_booleans,
-            &verifier_key_booleans,
-        )
-        .unwrap();
+        let d_delta =
+            MaskedWitnessVar::compute_d_delta(&witness_commitment_booleans, &verifier_key_booleans)
+                .unwrap();
 
         // Verify the result
         assert_eq!(d_delta.len(), 2);
@@ -410,28 +212,11 @@ mod revised_tests {
         ]];
 
         // Compute masked witnesses
-        let masked_witnesses = MaskedWitnessVarRevised::compute_masked_witness(
-            &witness_voles_booleans,
-            &d_delta_booleans,
-        )
-        .unwrap();
+        let masked_witnesses =
+            MaskedWitnessVar::compute_masked_witness(&witness_voles_booleans, &d_delta_booleans)
+                .unwrap();
 
         // Verify the result
         assert_eq!(masked_witnesses.len(), 1);
-    }
-
-    #[test]
-    fn test_combine() {
-        // Create test mask voles booleans
-        let mask_voles_booleans = vec![
-            vec![Boolean::constant(true), Boolean::constant(false)],
-            vec![Boolean::constant(false), Boolean::constant(true)],
-        ];
-
-        // Compute validation mask
-        let validation_mask = MaskedWitnessVarRevised::combine(&mask_voles_booleans).unwrap();
-
-        // Verify the result
-        assert!(validation_mask.len() > 0);
     }
 }
