+++ schmivitz-snark/src/gadgets/mod.rs
@@ -1,3 +1,5 @@
 pub mod masked_witness;
+pub mod validation;
 
-pub use masked_witness::MaskedWitnessVarRevised;
+pub use masked_witness::MaskedWitnessVar;
+pub use validation::ValidationVar;
