--- schmivitz-snark/src/tests/mod.rs
@@ -1,2 +1,0 @@
-pub mod circuit_constraint_tests;
-pub mod witness_validation_tests;
