+++ schmivitz/src/lib.rs
@@ -22,7 +22,7 @@ pub mod proof;
 // pub mod prove_from_bristol;
 mod vole;
 
-pub use proof::{to_serializable_proof, Proof, SerializableDecommitment, SerializableProof};
+pub use proof::Proof;
 // pub use prove_from_bristol::{cli_main, prove_and_verify_bristol, prove_sieve, verify_sieve};
 pub use vole::insecure;
 pub use vole::RandomVole;
